refactor: reorganize time_scheme header file into enum directory
Refs: #3
