+++ src/Wave.hpp
@@ -34,9 +34,9 @@
 #include <cstdlib>
 #include <filesystem>
 
+#include "enum/time_scheme.hpp"
 #include "parameters.hpp"
 #include "time_integrator.hpp"
-#include "time_scheme.hpp"
 
 using namespace dealii;
